[Feat] Add multilingual support using next-intl
